requirements.txt
matplotlib>=3.4
scikit-learn>=0.24
scipy>=1.7
numpy>=1.20
pillow>=9.0
torchsampler